Move creation of a brand new prf submission into new formio utility function for use across rebate years
