+++ app/server/app/utilities/formio.js
@@ -6,7 +6,10 @@ const {
   submissionPeriodOpen,
   formioCSBMetadata,
 } = require("../config/formio");
-const { checkFormSubmissionPeriodAndBapStatus } = require("./bap");
+const {
+  getBapDataForPRF,
+  checkFormSubmissionPeriodAndBapStatus,
+} = require("../utilities/bap");
 const log = require("./logger");
 
 /**
@@ -19,6 +22,110 @@ function getComboKeyFieldName({ rebateYear }) {
     : rebateYear === "2023"
     ? "_bap_entity_combo_key"
     : "";
+}
+
+/**
+ * @param {Object} param
+ * @param {'2022' | '2023'} param.rebateYear
+ * @param {express.Request} param.req
+ * @param {express.Response} param.res
+ * @returns {Promise<{ data: Object, metadata: Object, state: 'draft' }>}
+ */
+function fetchDataForPRFSubmission({ rebateYear, req, res }) {
+  /** @type {{
+   *  email: string
+   *  title: string
+   *  name: string
+   *  entity: import('./bap.js').BapSamEntity
+   *  comboKey: ?string
+   *  rebateId: ?string
+   *  frfReviewItemId: ?string
+   *  frfFormModified: ?string
+   * }} */
+  const {
+    email,
+    title,
+    name,
+    entity,
+    comboKey,
+    rebateId,
+    frfReviewItemId,
+    frfFormModified,
+  } = req.body;
+
+  return getBapDataForPRF(req, frfReviewItemId)
+    .then(({ frfRecordQuery, busRecordsQuery }) => {
+      const {
+        CSB_NCES_ID__c,
+        Primary_Applicant__r,
+        Alternate_Applicant__r,
+        Applicant_Organization__r,
+        CSB_School_District__r,
+        Fleet_Name__c,
+        School_District_Prioritized__c,
+        Total_Rebate_Funds_Requested__c,
+        Total_Infrastructure_Funds__c,
+      } = frfRecordQuery[0];
+
+      const busInfo = busRecordsQuery.map((record) => ({
+        busNum: record.Rebate_Item_num__c,
+        oldBusNcesDistrictId: CSB_NCES_ID__c,
+        oldBusVin: record.CSB_VIN__c,
+        oldBusModelYear: record.CSB_Model_Year__c,
+        oldBusFuelType: record.CSB_Fuel_Type__c,
+        newBusFuelType: record.CSB_Replacement_Fuel_Type__c,
+        hidden_bap_max_rebate: record.CSB_Funds_Requested__c,
+      }));
+
+      /**
+       * NOTE: `purchaseOrders` is initialized as an empty array to fix some
+       * issue with the field being changed to an object when the form loads
+       */
+      const submission = {
+        data: {
+          bap_hidden_entity_combo_key: comboKey,
+          hidden_application_form_modified: frfFormModified,
+          hidden_current_user_email: email,
+          hidden_current_user_title: title,
+          hidden_current_user_name: name,
+          hidden_sam_uei: entity.UNIQUE_ENTITY_ID__c,
+          hidden_sam_efti: entity.ENTITY_EFT_INDICATOR__c || "0000",
+          hidden_sam_elec_bus_poc_email: entity.ELEC_BUS_POC_EMAIL__c,
+          hidden_sam_alt_elec_bus_poc_email: entity.ALT_ELEC_BUS_POC_EMAIL__c,
+          hidden_sam_govt_bus_poc_email: entity.GOVT_BUS_POC_EMAIL__c,
+          hidden_sam_alt_govt_bus_poc_email: entity.ALT_GOVT_BUS_POC_EMAIL__c,
+          hidden_bap_rebate_id: rebateId,
+          hidden_bap_district_id: CSB_NCES_ID__c,
+          hidden_bap_primary_name: Primary_Applicant__r?.Name,
+          hidden_bap_primary_title: Primary_Applicant__r?.Title,
+          hidden_bap_primary_phone_number: Primary_Applicant__r?.Phone,
+          hidden_bap_primary_email: Primary_Applicant__r?.Email,
+          hidden_bap_alternate_name: Alternate_Applicant__r?.Name || "",
+          hidden_bap_alternate_title: Alternate_Applicant__r?.Title || "",
+          hidden_bap_alternate_phone_number: Alternate_Applicant__r?.Phone || "", // prettier-ignore
+          hidden_bap_alternate_email: Alternate_Applicant__r?.Email || "",
+          hidden_bap_org_name: Applicant_Organization__r?.Name,
+          hidden_bap_district_name: CSB_School_District__r?.Name,
+          hidden_bap_fleet_name: Fleet_Name__c,
+          hidden_bap_prioritized: School_District_Prioritized__c,
+          hidden_bap_requested_funds: Total_Rebate_Funds_Requested__c,
+          hidden_bap_infra_max_rebate: Total_Infrastructure_Funds__c,
+          busInfo,
+          purchaseOrders: [],
+        },
+        /** Add custom metadata to track formio submissions from wrapper. */
+        metadata: { ...formioCSBMetadata },
+        state: "draft",
+      };
+
+      return submission;
+    })
+    .catch((error) => {
+      // NOTE: logged in bap verifyBapConnection
+      const errorStatus = 500;
+      const errorMessage = `Error getting data for a new Payment Request form submission from the BAP.`;
+      return res.status(errorStatus).json({ message: errorMessage });
+    });
 }
 
 /**
@@ -391,6 +498,56 @@ function fetchPRFSubmissions({ rebateYear, req, res }) {
     });
 }
 
+/**
+ * @param {Object} param
+ * @param {'2022' | '2023'} param.rebateYear
+ * @param {express.Request} param.req
+ * @param {express.Response} param.res
+ */
+function createPRFSubmission({ rebateYear, req, res }) {
+  const { bapComboKeys, body } = req;
+  const { mail } = req.user;
+  const { comboKey } = body;
+
+  const formioFormUrl = formUrl[rebateYear].prf;
+
+  if (!formioFormUrl) {
+    const errorStatus = 400;
+    const errorMessage = `Formio form URL does not exist for ${rebateYear} PRF.`;
+    return res.status(errorStatus).json({ message: errorMessage });
+  }
+
+  if (!submissionPeriodOpen[rebateYear].prf) {
+    const errorStatus = 400;
+    const errorMessage = `${rebateYear} CSB Payment Request form enrollment period is closed.`;
+    return res.status(errorStatus).json({ message: errorMessage });
+  }
+
+  if (!bapComboKeys.includes(comboKey)) {
+    const logMessage =
+      `User with email '${mail}' attempted to post a new ${rebateYear} ` +
+      `PRF submission without a matching BAP combo key.`;
+    log({ level: "error", message: logMessage, req });
+
+    const errorStatus = 401;
+    const errorMessage = `Unauthorized.`;
+    return res.status(errorStatus).json({ message: errorMessage });
+  }
+
+  fetchDataForPRFSubmission({ rebateYear, req, res }).then((submission) => {
+    axiosFormio(req)
+      .post(`${formioFormUrl}/submission`, submission)
+      .then((axiosRes) => axiosRes.data)
+      .then((submission) => res.json(submission))
+      .catch((error) => {
+        // NOTE: error is logged in axiosFormio response interceptor
+        const errorStatus = error.response?.status || 500;
+        const errorMessage = `Error posting Formio ${rebateYear} Payment Request form submission.`;
+        return res.status(errorStatus).json({ message: errorMessage });
+      });
+  });
+}
+
 module.exports = {
   uploadS3FileMetadata,
   downloadS3FileMetadata,
@@ -401,4 +558,5 @@ module.exports = {
   updateFRFSubmission,
   //
   fetchPRFSubmissions,
+  createPRFSubmission,
 };
